+++ packages/path_provider/path_provider_foundation/darwin/Classes/PathProviderPlugin.swift
@@ -24,12 +24,19 @@ public class PathProviderPlugin: NSObject, FlutterPlugin, PathProviderApi {
 
   func getDirectoryPath(type: DirectoryType) -> String? {
     var path = getDirectory(ofType: fileManagerDirectoryForType(type))
+  #if os(macOS)
+    // In a non-sandboxed app, this is a shared directory where applications are
+    // expected to use its bundle ID as a subdirectory. (For non-sandboxed apps,
+    // adding the extra path is harmless).
+    // This is not done for iOS, for compatibility with older versions of the
+    // plugin.
     if type == .applicationSupport {
       if let basePath = path {
         let basePathURL = URL.init(fileURLWithPath: basePath)
         path = basePathURL.appendingPathComponent(Bundle.main.bundleIdentifier!).path
       }
     }
+  #endif
     return path
   }
 }