+++ packages/file_selector/file_selector/CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+* Updates example code for `use_build_context_synchronously` lint.
+* Updates minimum Flutter version to 3.0.
+
 ## 0.9.2+2
 
 * Improves API docs and examples.