+++ packages/in_app_purchase/in_app_purchase/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.1.2
+
+* Updates example code for `use_build_context_synchronously` lint.
+* Updates minimum Flutter version to 3.0.
+
 ## 3.1.1
 
 * Adds screenshots to pubspec.yaml.